refactor!: move downloading package from remote or load from cache to Environment
