+++ vrc-get-vpm/src/environment.rs
@@ -5,24 +5,29 @@ use crate::repository::local::LocalCachedRepository;
 use crate::repository::{RemotePackages, RemoteRepository};
 use crate::structs::package::PackageJson;
 use crate::structs::setting::UserRepoSetting;
-use crate::traits::PackageCollection;
-use crate::utils::{JsonMapExt, PathBufExt};
+use crate::traits::{PackageCollection, RemotePackageDownloader};
+use crate::utils::{JsonMapExt, MapResultExt, PathBufExt, Sha256AsyncWrite};
 use crate::version::{UnityVersion, Version, VersionRange};
 use crate::{
     is_truthy, load_json_or_default, to_json_vec, unity_compatible, PackageInfo,
     PreDefinedRepoSource, RepoSource, DEFINED_REPO_SOURCES,
 };
 use futures::future::join_all;
+use futures::prelude::*;
+use futures::StreamExt;
+use hex::FromHex;
 use indexmap::IndexMap;
 use itertools::Itertools;
-use reqwest::{Client, Url};
+use reqwest::{Client, Response, Url};
 use serde_json::{from_value, to_value, Map, Value};
 use std::cmp::Reverse;
 use std::collections::{HashMap, HashSet};
+use std::io::SeekFrom;
 use std::path::{Path, PathBuf};
 use std::{env, fmt, io};
-use tokio::fs::{create_dir_all, remove_file, File};
-use tokio::io::AsyncWriteExt;
+use tokio::fs::{create_dir_all, remove_file, File, OpenOptions};
+use tokio::io::{AsyncReadExt, AsyncSeekExt, AsyncWriteExt};
+use tokio_util::compat::*;
 
 pub(crate) use crate::environment::uesr_package_collection::UserPackageCollection;
 
@@ -452,6 +457,151 @@ impl Environment {
         self.settings_changed = false;
         Ok(())
     }
+}
+
+impl RemotePackageDownloader for Environment {
+    async fn get_package(
+        &self,
+        repository: &LocalCachedRepository,
+        package: &PackageJson,
+    ) -> io::Result<File> {
+        let zip_file_name = format!("vrc-get-{}-{}.zip", &package.name, &package.version);
+        let zip_path = self
+            .global_dir
+            .to_owned()
+            .joined("Repos")
+            .joined(&package.name)
+            .joined(&zip_file_name);
+        let sha_path = zip_path.with_extension("zip.sha256");
+
+        if let Some(cache_file) = try_load_package_cache(&zip_path, &sha_path, None).await {
+            Ok(cache_file)
+        } else {
+            create_dir_all(zip_path.parent().unwrap()).await?;
+
+            Ok(download_package_zip(
+                self.http.as_ref(),
+                repository.headers(),
+                &zip_path,
+                &sha_path,
+                &zip_file_name,
+                &package.url,
+            )
+            .await?)
+        }
+    }
+}
+
+/// Try to load from the zip file
+///
+/// # Arguments
+///
+/// * `zip_path`: the path to zip file
+/// * `sha_path`: the path to sha256 file
+/// * `sha256`: sha256 hash if specified
+///
+/// returns: Option<File> readable zip file file or None
+async fn try_load_package_cache(
+    zip_path: &Path,
+    sha_path: &Path,
+    sha256: Option<&str>,
+) -> Option<File> {
+    let mut cache_file = File::open(zip_path).await.ok()?;
+
+    let mut buf = [0u8; 256 / 4];
+    File::open(sha_path)
+        .await
+        .ok()?
+        .read_exact(&mut buf)
+        .await
+        .ok()?;
+
+    let hex: [u8; 256 / 8] = FromHex::from_hex(buf).ok()?;
+
+    // is stored sha doesn't match sha in repo: current cache is invalid
+    if let Some(repo_hash) = sha256.and_then(|x| <[u8; 256 / 8] as FromHex>::from_hex(x).ok()) {
+        if repo_hash != hex {
+            return None;
+        }
+    }
+
+    let mut hasher = Sha256AsyncWrite::new(tokio::io::sink());
+
+    tokio::io::copy(&mut cache_file, &mut hasher).await.ok()?;
+
+    let hash = &hasher.finalize().1[..];
+    if hash != &hex[..] {
+        return None;
+    }
+
+    cache_file.seek(SeekFrom::Start(0)).await.ok()?;
+
+    Some(cache_file)
+}
+
+/// downloads the zip file from the url to the specified path
+///
+/// # Arguments
+///
+/// * `http`: http client. returns error if none
+/// * `zip_path`: the path to zip file
+/// * `sha_path`: the path to sha256 file
+/// * `zip_file_name`: the name of zip file. will be used in the sha file
+/// * `url`: url to zip file
+///
+/// returns: Result<File, Error> the readable zip file.
+async fn download_package_zip(
+    http: Option<&Client>,
+    headers: &IndexMap<String, String>,
+    zip_path: &Path,
+    sha_path: &Path,
+    zip_file_name: &str,
+    url: &str,
+) -> io::Result<File> {
+    let Some(http) = http else {
+        return Err(io::Error::new(io::ErrorKind::NotFound, "Offline mode"));
+    };
+
+    // file not found: err
+    let cache_file = OpenOptions::new()
+        .read(true)
+        .write(true)
+        .create(true)
+        .open(&zip_path)
+        .await?;
+
+    let mut request = http.get(url);
+
+    for (name, header) in headers {
+        request = request.header(name, header);
+    }
+
+    let mut response = request
+        .send()
+        .await
+        .and_then(Response::error_for_status)
+        .err_mapped()?
+        .bytes_stream()
+        .map(|x| x.err_mapped())
+        .into_async_read()
+        .compat();
+
+    let mut writer = Sha256AsyncWrite::new(cache_file);
+    tokio::io::copy(&mut response, &mut writer).await?;
+
+    let (mut cache_file, hash) = writer.finalize();
+
+    cache_file.flush().await?;
+    cache_file.seek(SeekFrom::Start(0)).await?;
+
+    // write sha file
+    tokio::fs::write(
+        &sha_path,
+        format!("{} {}\n", hex::encode(&hash[..]), zip_file_name),
+    )
+    .await?;
+
+    Ok(cache_file)
 }
 
 #[derive(Clone, Copy)]
